+++ src/main/java/ortus/boxlang/lsp/LanguageServer.java
@@ -41,6 +41,8 @@ public class LanguageServer implements org.eclipse.lsp4j.services.LanguageServer
 			capabilities.setDocumentSymbolProvider( true );
 			capabilities.setDefinitionProvider( true );
 			CompletionOptions completionOptions = new CompletionOptions();
+			capabilities.setReferencesProvider( true );
+
 			completionOptions.setTriggerCharacters( List.of( "." ) );
 			// completionOptions.
 			capabilities.setCompletionProvider( completionOptions );